+++ src/App.tsx
@@ -15,7 +15,7 @@ import ExportImage from 'screens/internal/ShareImage/ChartExportImage';
 import ShareImage from 'screens/internal/ShareImage/ShareImage';
 import AlertUnsubscribe from 'screens/AlertUnsubscribe/AlertUnsubscribe';
 import NavBar from 'components/AppBar';
-import Footer from 'components/NewFooter';
+import Footer from 'components/Footer';
 import ScrollToTop from 'components/ScrollToTop';
 import theme from 'assets/theme';
 import { getFeedbackSurveyUrl } from 'components/Banner';
