+++ src/components/AppBar/DonateButton/index.ts
@@ -1,0 +1,3 @@
+import { DonateButton, DonateButtonHeart } from './DonateButton';
+
+export { DonateButton, DonateButtonHeart };
