--- src/components/AppBar/MenuButton.style.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-import { BaseButton } from 'components/Button';
-
-export const StyledButton = styled(BaseButton)`
-  text-transform: none;
-  font-size: 16px;
-`;
