--- src/components/AppBar/MenuButton.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { StyledButton } from './MenuButton.style';
-
-type MenuButtonProps = React.ComponentProps<typeof StyledButton>;
-
-const MenuButton: React.FC<MenuButtonProps> = props => {
-  return <StyledButton {...props} />;
-};
-
-export default MenuButton;
