+++ src/components/MenuContent/FeaturedSection.tsx
