+++ src/components/MenuContent/Menu.style.tsx
@@ -1,51 +1,12 @@
 import styled, { css } from 'styled-components';
 import { Link } from 'react-router-dom';
 import { COLOR_MAP } from 'common/colors';
-import { mobileBreakpoint, materialSMBreakpoint } from 'assets/theme/sizes';
+import { mobileBreakpoint } from 'assets/theme/sizes';
 import ArrowForwardIcon from '@material-ui/icons/ArrowForward';
 import LinkButton from 'components/LinkButton';
+import fonts from 'common/theme/fonts';
 
-const mobileBreakpointPlus = '960px';
-
-export const StyledFooter = styled.footer`
-  box-sizing: content-box;
-  background: ${COLOR_MAP.BLACK};
-  padding: 2rem 1.5rem;
-
-  @media (min-width: ${mobileBreakpoint}) {
-    padding: 3rem;
-  }
-
-  @media (min-width: ${mobileBreakpointPlus}) {
-    padding: 4rem;
-  }
-`;
-
-// Need to add a good amount of bottom padding for mobile screens to make sure the bottom
-// of the mega menu isn't covered by the browser's menu bar
-export const StyledMegaMenu = styled.nav`
-  background: white;
-  border-top: 1px solid ${COLOR_MAP.GREY_2};
-  border-bottom: 1px solid ${COLOR_MAP.GREY_2};
-  transform: translateY(64px);
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  padding: 2rem 1.5rem 8rem;
-  height: calc(100vh - 224px);
-  overflow-y: auto;
-  box-sizing: content-box;
-  box-shadow: 0px 15px 30px -15px rgba(0, 0, 0, 0.2);
-
-  @media (min-width: ${materialSMBreakpoint}) {
-    height: auto;
-  }
-
-  @media (min-width: ${mobileBreakpoint}) {
-    padding: 3rem;
-  }
-`;
+export const mobileBreakpointPlus = '960px';
 
 export const ContentWrapper = styled.div`
   display: flex;
@@ -102,7 +63,7 @@ export const Section = styled.div`
 `;
 
 export const SectionHeader = styled.h2<{ $desktopOnly?: boolean }>`
-  ${props => props.theme.fonts.regularBookMidWeight};
+  ${fonts.regularBookMidWeight};
   color: ${props => props.theme.palette.megaMenu.gray};
   text-transform: uppercase;
   margin: 0 0 1.25rem;
@@ -131,15 +92,8 @@ export const Row = styled.div`
   margin-bottom: 0.25rem;
 `;
 
-export const RowWithSpacing = styled(Row)`
-  margin-bottom: 0;
-  &:not(:last-of-type) {
-    margin-bottom: 1.75rem;
-  }
-`;
-
 export const BodyCopy = css`
-  ${props => props.theme.fonts.regularBook};
+  ${fonts.regularBook};
   line-height: 1.4;
   margin: 0;
 `;
@@ -152,7 +106,7 @@ export const FeaturedDescription = styled.p`
 
 export const AboutCopy = styled.p`
   ${BodyCopy};
-  color: ${props => props.theme.palette.megaMenu.primaryText};
+  color: ${props => props.theme.palette.megaMenu.secondaryText};
   margin-bottom: 2rem;
   text-align: ${props => props.theme.palette.megaMenu.aboutUsTextAlignment};
 
@@ -162,7 +116,7 @@ export const AboutCopy = styled.p`
 `;
 
 export const ButtonBase = css`
-  ${props => props.theme.fonts.regularBookMidWeight};
+  ${fonts.regularBookMidWeight};
   width: fit-content;
   text-transform: none;
   line-height: 1.4;
@@ -170,6 +124,7 @@ export const ButtonBase = css`
 
 export const TextLink = styled(Link)`
   ${ButtonBase};
+  ${fonts.regularBookMidWeight};
 
   color: ${props => props.theme.palette.megaMenu.primaryText};
 
@@ -201,16 +156,15 @@ export const LearnLink = styled(TextLink)`
   }
 `;
 
-export const OutlinedButton = styled(LinkButton)<{ desktopOnly?: boolean }>`
+export const OutlinedButton = styled(LinkButton)`
   ${ButtonBase};
-
   color: ${props => props.theme.palette.megaMenu.buttonContent};
-
   font-size: 0.8725rem;
   padding: 0.4rem 0.75rem;
   border: ${props => `1px solid ${props.theme.palette.megaMenu.buttonBorder}`};
-  display: ${({ desktopOnly }) => desktopOnly && 'none'};
   white-space: nowrap;
+  margin-top: 2rem;
+
   &:nth-child(2) {
     margin-left: 0.75rem;
   }
@@ -222,7 +176,18 @@ export const OutlinedButton = styled(LinkButton)<{ desktopOnly?: boolean }>`
   }
 
   @media (min-width: ${mobileBreakpoint}) {
-    display: inherit;
+    margin-top: 0;
+  }
+`;
+
+export const RowWithSpacing = styled(Row)`
+  margin-bottom: 0;
+  &:not(:last-of-type) {
+    margin-bottom: 1.75rem;
+  }
+
+  ${OutlinedButton} {
+    margin-top: 0;
   }
 `;
 
@@ -263,4 +228,5 @@ export const LogoWrapper = styled(Link)`
 
 export const NonWrappingSpan = styled.span`
   white-space: nowrap;
+  ${fonts.regularBookMidWeight};
 `;
