+++ src/components/MenuContent/SocialButtonsBlock.tsx
