+++ src/components/MenuContent/TextAndIconWithSpecialWrapping.tsx
