+++ src/components/MenuContent/index.ts
@@ -1,0 +1,3 @@
+import MenuContent from './MenuContent';
+
+export default MenuContent;
